+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* [#13](https://github.com/sdss/lvmgort/issues/13) Adds a signal handler for `SIGINT` and `SIGTERM`. When `GortObserver` is running and an interrupt is received the signal handler will run the cleanup routine.
+
+
 ## 0.7.1 - December 20, 2023
 
 ### ✨ Improved
 
-* [#13](https://github.com/sdss/lvmgort/issues/13) Adds a signal handler for `SIGINT` and `SIGTERM`. When `GortObserver` is running and an interrupt is received the signal handler will run the cleanup routine.
 * [#14](https://github.com/sdss/lvmgort/issues/14) Turn off dome lights as part of the `startup` and `cleanup` recipes.
 * Use `lvmnps all-off`` to turn off all lamps faster.